move some monitoring constant and function to common/monitoring package
